poprawki w cenach, przerabianie na int z double (grosze dla biblioteki phpmoney)

+++ database/migrations/2021_11_18_133341_create_orders_table.php
@@ -21,7 +21,7 @@ class CreateOrdersTable extends Migration
             $table->enum('section',['small_ads','cars','estates','job','services','truck'])->comment('sekcja której dotyczy płatność');
             $table->string('users_id');
             $table->enum('status',['paid','in progres','unpaid','canceled'])->default('unpaid')->comment('unpaid');
-            $table->double('summary_price',2)->comment('unpaid');
+            $table->integer('summary_price')->comment('unpaid');
             
             $table->timestamps();
